# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/candidates": {
      "post": {
        "operationId": "CandidatesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CandidateResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create candidate profile",
        "tags": [
          "Candidates"
        ]
      },
      "get": {
        "operationId": "CandidatesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "schema": {
              "default": "created_at",
              "example": "created_at",
              "type": "string",
              "enum": [
                "created_at",
                "updated_at",
                "first_name",
                "last_name",
                "email",
                "current_location",
                "total_experience_years",
                "is_active"
              ]
            }
          },
          {
            "name": "sort_order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor-based pagination. ISO timestamp for created_at. When provided, cursor pagination is used and `sort_by` must be `created_at`.",
            "schema": {
              "example": "2026-01-15T10:30:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "first_name",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Sakshi",
              "type": "string"
            }
          },
          {
            "name": "last_name",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Sharma",
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "example": "sakshi@example.com",
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "schema": {
              "example": "+919876543210",
              "type": "string"
            }
          },
          {
            "name": "source_type",
            "required": false,
            "in": "query",
            "schema": {
              "example": "LINKEDIN",
              "type": "string",
              "enum": [
                "REFERRAL",
                "LINKEDIN",
                "COMPANY_PORTAL",
                "JOB_BOARD",
                "AGENCY",
                "WALK_IN",
                "OTHER"
              ]
            }
          },
          {
            "name": "current_location",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Bengaluru",
              "type": "string"
            }
          },
          {
            "name": "total_experience_years",
            "required": false,
            "in": "query",
            "schema": {
              "example": 5.5,
              "type": "number"
            }
          },
          {
            "name": "is_active",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          },
          {
            "name": "skill_id",
            "required": false,
            "in": "query",
            "description": "Filter candidates by skill UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidates fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/PaginatedResponseDto"
                        },
                        {
                          "properties": {
                            "data": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/CandidateResponseDto"
                              }
                            }
                          }
                        }
                      ]
                    },
                    {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/BaseResponseDto"
                        },
                        {
                          "properties": {
                            "data": {
                              "type": "object",
                              "properties": {
                                "data": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/CandidateResponseDto"
                                  }
                                },
                                "limit": {
                                  "type": "number",
                                  "example": 10
                                },
                                "next_cursor": {
                                  "type": "string",
                                  "nullable": true,
                                  "example": "2026-01-15T10:30:00.000Z"
                                },
                                "has_more": {
                                  "type": "boolean",
                                  "example": true
                                }
                              }
                            }
                          }
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List candidates (offset or cursor pagination)",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/api/candidates/{id}": {
      "patch": {
        "operationId": "CandidatesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Candidate UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CandidateResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update candidate profile",
        "tags": [
          "Candidates"
        ]
      },
      "get": {
        "operationId": "CandidatesController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Candidate UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CandidateResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get candidate by id",
        "tags": [
          "Candidates"
        ]
      },
      "delete": {
        "operationId": "CandidatesController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Candidate UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SoftDeleteCandidateResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Soft delete candidate",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/api/candidates/{id}/skills": {
      "put": {
        "operationId": "CandidatesController_upsertSkills",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Candidate UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertCandidateSkillsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate skills updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CandidateSkillResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Add/update candidate skills",
        "tags": [
          "Candidates"
        ]
      },
      "get": {
        "operationId": "CandidatesController_listSkills",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Candidate UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate skills fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CandidateSkillResponseDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get candidate skills",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/api/candidates/{id}/skills/{skillId}": {
      "delete": {
        "operationId": "CandidatesController_removeSkill",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Candidate UUID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skillId",
            "required": true,
            "in": "path",
            "description": "Skill UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate skill removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/RemoveCandidateSkillResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remove candidate skill",
        "tags": [
          "Candidates"
        ]
      }
    },
    "/api/candidates/{id}/documents/upload": {
      "post": {
        "operationId": "CandidatesController_uploadDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Candidate UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "document_type",
                  "file"
                ],
                "properties": {
                  "document_type": {
                    "type": "string",
                    "enum": [
                      "RESUME",
                      "COVER_LETTER",
                      "PORTFOLIO",
                      "CERTIFICATION",
                      "ID_PROOF",
                      "OTHER"
                    ],
                    "example": "RESUME"
                  },
                  "is_latest": {
                    "type": "boolean",
                    "example": true,
                    "description": "Only applicable for RESUME documents"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate document uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CandidateDocumentResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Upload candidate document (stores file + metadata)",
        "tags": [
          "Candidates"
        ]
      }
    },
...
  },
  "info": {
    "title": "Recruitment Platform API",
    "description": "Backend APIs for Recruitment Pipeline & Interview Intelligence Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "BaseResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Request processed successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Records fetched successfully"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "success",
          "message",
          "pagination"
        ]
      },
...
      "SoftDeleteCandidateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b9b5d98d-4f82-4d87-8a84-0f4f6c9dfef2"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "UpsertCandidateSkillsDto": {
        "type": "object",
        "properties": {
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateSkillInputDto"
            }
          }
        },
        "required": [
          "skills"
        ]
      },
      "RemoveCandidateSkillResponseDto": {
        "type": "object",
        "properties": {
          "candidate_id": {
            "type": "string",
            "description": "Candidate UUID"
          },
          "skill_id": {
            "type": "string",
            "description": "Skill UUID"
          }
        },
        "required": [
          "candidate_id",
          "skill_id"
        ]
      },
...
    }
  }
}